3.9 Выносит утилитарные функции в отдельный файл
